refactor: change tenant and user ID types from int to int64 for improved compatibility

+++ auth/tenant_id.go
@@ -6,12 +6,12 @@ import (
 
 type tenantIDKey struct{}
 
-func WithTenantID(ctx context.Context, tenantID int) context.Context {
+func WithTenantID(ctx context.Context, tenantID int64) context.Context {
 	return context.WithValue(ctx, tenantIDKey{}, tenantID)
 }
 
-func TenantIDFrom(ctx context.Context) int {
-	tenantID, ok := ctx.Value(tenantIDKey{}).(int)
+func TenantIDFrom(ctx context.Context) int64 {
+	tenantID, ok := ctx.Value(tenantIDKey{}).(int64)
 	if !ok {
 		return 0
 	}
